RN-685: Switched to using MockTupaiaApiClient in report-server tests
