+++ packages/report-server/src/__tests__/reportBuilder/customReport/testCustomReport.test.ts
@@ -4,11 +4,10 @@
  */
 
 import { AccessPolicy } from '@tupaia/access-policy';
+import { MockEntityApi, MockTupaiaApiClient } from '@tupaia/api-client';
 import { ReqContext } from '../../../reportBuilder/context';
 
 import { testCustomReport } from '../../../reportBuilder/customReports/testCustomReport';
-
-import { entityApiMock } from '../testUtils';
 
 describe('testCustomReport', () => {
   const HIERARCHY = 'test_hierarchy';
@@ -44,14 +43,10 @@ describe('testCustomReport', () => {
     ],
   };
 
-  const apiMock = entityApiMock(ENTITIES, RELATIONS);
-
   const reqContext: ReqContext = {
     hierarchy: HIERARCHY,
     permissionGroup: 'Public',
-    services: {
-      entity: apiMock,
-    } as ReqContext['services'],
+    services: new MockTupaiaApiClient({ entity: new MockEntityApi(ENTITIES, RELATIONS) }),
     accessPolicy: new AccessPolicy({ AU: ['Public'] }),
   };
