--- packages/report-server/src/__tests__/reportBuilder/testUtils/index.ts
@@ -1,6 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
- */
-
-export { entityApiMock } from './entityApiMock';
